### 2021-08-09 #3592 @whedon
[PRE REVIEW]: Feature-engine: A Python package for feature engineering for machine learning
**Submitting author:** @solegalli (<a href="http://orcid.org/0000-0003-3734-8628">Soledad</a>)
**Repository:** <a href="https://github.com/feature-engine/feature_engine" target ="_blank">https://github.com/feature-engine/feature_engine</a>
**Version:** v1.1.1
