### 2021-08-09 #3592 @arfon
:wave: @solegalli  – thanks for your submission to JOSS. We're currently managing a large backlog of submissions and the editor most appropriate for your area is already rather busy.

In the interim though, please updated your reference keys so they match between your Markdown doc and references (e.g., change `@pandas:2010` in your `paper.md` to `@pandas` in `paper.bib`).

Also, could you explain why there is only a single author for this paper when there look to have been significant contributions made by others? Has the authorship of this paper been discussed anywhere before submission?
