### 2021-08-09 6c15863
adds draft citations, completes about, reorganises logos location
