### 2021-08-09 6647f0a
update format inline citations
