### 2021-08-09 #3592 @solegalli
Hi @arfon thanks a lot for starting the process.

The reference keys were updated as requested.

At the moment, I am the founder, main developer and maintainer of the project. More details about the governance can be found [here](https://feature-engine.readthedocs.io/en/latest/governance.html).

The community is growing and we have growing support from users, fortunately, which is quite encouraging, but we are not at a stage yet, where we have a committee who makes joint decisions.

In the submitted article, I followed the initiative of 2 other Python open source software libraries for machine learning, which, while had support from the community at time of publication, the authors of the article where the founders and maintainers of the project. See for example [yellowbricks](https://joss.theoj.org/papers/10.21105/joss.01075) and [mlxtend](https://www.theoj.org/joss-papers/joss.00638/10.21105.joss.00638.pdf).

I cc here @ChristopherGS who has supported the project from the beginning as explained in the acknowledgements section.

Thanks a lot
