### 2021-08-10 681bc8a
Simplify `get_feature_names`
- store input features on fit
- use transformed array to compare output of `get_feature_names` in tests
