+++ tests/test_encoding/test_count_frequency_encoder.py
@@ -44,7 +44,7 @@ def test_encode_1_variable_with_counts(df_enc):
     assert encoder.n_features_in_ == 3
     # transform params
     pd.testing.assert_frame_equal(X, transf_df)
-    assert encoder.get_feature_names() == ["var_A"]
+    assert encoder.get_feature_names() == X.columns.tolist()
 
 
 def test_automatically_select_variables_encode_with_frequency(df_enc):
@@ -111,7 +111,7 @@ def test_automatically_select_variables_encode_with_frequency(df_enc):
     assert encoder.n_features_in_ == 3
     # transform params
     pd.testing.assert_frame_equal(X, transf_df)
-    assert encoder.get_feature_names() == ["var_A", "var_B"]
+    assert encoder.get_feature_names() == X.columns.tolist()
 
 
 def test_error_if_encoding_method_not_permitted_value():
@@ -176,7 +176,6 @@ def test_ignore_variable_format_with_frequency(df_vartypes):
     assert encoder.n_features_in_ == 5
     # transform params
     pd.testing.assert_frame_equal(X, transf_df)
-    assert encoder.get_feature_names() == ["Name", "City", "Age", "Marks", "dob"]
     assert encoder.get_feature_names() == X.columns.tolist()
 
 
@@ -205,7 +204,7 @@ def test_column_names_are_numbers(df_numeric_columns):
     assert encoder.n_features_in_ == 5
     # transform params
     pd.testing.assert_frame_equal(X, transf_df)
-    assert encoder.get_feature_names() == [0, 1, 2, 3]
+    assert encoder.get_feature_names() == X.columns.tolist()
 
 
 def test_variables_cast_as_category(df_enc_category_dtypes):
@@ -242,5 +241,5 @@ def test_variables_cast_as_category(df_enc_category_dtypes):
 
     encoder = CountFrequencyEncoder(encoding_method="frequency", variables=["var_A"])
     X = encoder.fit_transform(df_enc_category_dtypes)
+    assert encoder.get_feature_names() == X.columns.tolist()
     assert X["var_A"].dtypes == float
-    assert encoder.get_feature_names() == ["var_A"]
