+++ tests/test_encoding/test_ordinal_encoder.py
@@ -24,7 +24,7 @@ def test_ordered_encoding_1_variable(df_enc):
     assert encoder.n_features_in_ == 2
     # test transform output
     pd.testing.assert_frame_equal(X, transf_df[["var_A", "var_B"]])
-    assert encoder.get_feature_names() == ["var_A"]
+    assert encoder.get_feature_names() == X.columns.tolist()
 
 
 def test_arbitrary_encoding_automatically_find_variables(df_enc):
@@ -49,7 +49,7 @@ def test_arbitrary_encoding_automatically_find_variables(df_enc):
     assert encoder.n_features_in_ == 3
     # test transform output
     pd.testing.assert_frame_equal(X, transf_df)
-    assert encoder.get_feature_names() == ["var_A", "var_B"]
+    assert encoder.get_feature_names() == X.columns.tolist()
 
 
 def test_error_if_encoding_method_not_allowed():
@@ -117,7 +117,7 @@ def test_ordered_encoding_1_variable_ignore_format(df_enc_numeric):
     assert encoder.n_features_in_ == 2
     # test transform output
     pd.testing.assert_frame_equal(X, transf_df[["var_A", "var_B"]])
-    assert encoder.get_feature_names() == ["var_A"]
+    assert encoder.get_feature_names() == X.columns.tolist()
 
 
 def test_arbitrary_encoding_automatically_find_variables_ignore_format(df_enc_numeric):
@@ -144,7 +144,6 @@ def test_arbitrary_encoding_automatically_find_variables_ignore_format(df_enc_nu
     assert encoder.n_features_in_ == 2
     # test transform output
     pd.testing.assert_frame_equal(X, transf_df)
-    assert encoder.get_feature_names() == ["var_A", "var_B"]
     assert encoder.get_feature_names() == X.columns.tolist()
 
 
@@ -160,5 +159,5 @@ def test_variables_cast_as_category(df_enc_category_dtypes):
 
     # test transform output
     pd.testing.assert_frame_equal(X, transf_df[["var_A", "var_B"]], check_dtype=False)
-    assert encoder.get_feature_names() == ["var_A"]
+    assert encoder.get_feature_names() == X.columns.tolist()
     assert X["var_A"].dtypes == int
