+++ feature_engine/encoding/base_encoder.py
@@ -1,5 +1,5 @@
 import warnings
-from typing import List, Optional, Union
+from typing import List, Union
 
 import pandas as pd
 from sklearn.base import BaseEstimator, TransformerMixin
@@ -58,6 +58,9 @@ class BaseCategoricalTransformer(BaseEstimator, TransformerMixin):
         else:
             # select all variables or check variables entered by the user
             self.variables_ = _find_all_variables(X, self.variables)
+
+        # save input features
+        self.input_features_: List[Union[str, int]] = X.columns.tolist()
 
         # check if dataset contains na
         _check_contains_na(X, self.variables_)
@@ -205,25 +208,13 @@ class BaseCategoricalTransformer(BaseEstimator, TransformerMixin):
         tags_dict["_xfail_checks"]["check_estimators_nan_inf"] = "transformer allows NA"
         return tags_dict
 
-    def get_feature_names(
-        self,
-        input_features: Optional[List[Union[str, int]]] = None,
-    ) -> List:
+    def get_feature_names(self) -> List:
         """
         Return feature names for output features.
-
-        Parameters
-        ----------
-        input_features : list of str of shape (n_features,)
-            String names for input features if available. By default,
-            self.variables_ is used
 
         -------
         feature_names : list of str of shape (n_output_features,) of feature names
         """
         check_is_fitted(self)
 
-        if input_features is None:
-            input_features = self.variables_
-
-        return [feature for feature in input_features if feature in self.encoder_dict_]
+        return self.input_features_
