### 2021-08-10 8db2aa1
Simplify `get_feature_names`
- store input features on fit
- use transformed array to compare output of `get_feature_names` in tests
