+++ tests/test_encoding/test_onehot_encoder.py
@@ -49,29 +49,6 @@ def test_encode_categories_in_k_binary_plus_select_vars_automatically(df_enc_big
     }
     # test transform output
     assert X.sum().to_dict() == transf
-    assert encoder.get_feature_names() == [
-        "var_A_A",
-        "var_A_B",
-        "var_A_C",
-        "var_A_D",
-        "var_A_E",
-        "var_A_F",
-        "var_A_G",
-        "var_B_A",
-        "var_B_B",
-        "var_B_C",
-        "var_B_D",
-        "var_B_E",
-        "var_B_F",
-        "var_B_G",
-        "var_C_A",
-        "var_C_B",
-        "var_C_C",
-        "var_C_D",
-        "var_C_E",
-        "var_C_F",
-        "var_C_G",
-    ]
     assert encoder.get_feature_names() == X.columns.tolist()
     assert "var_A" not in X.columns
 
@@ -114,20 +91,7 @@ def test_encode_categories_in_k_minus_1_binary_plus_list_of_variables(df_enc_big
     # test transform output
     for col in transf.keys():
         assert X[col].sum() == transf[col]
-    assert encoder.get_feature_names() == [
-        "var_A_A",
-        "var_A_B",
-        "var_A_C",
-        "var_A_D",
-        "var_A_E",
-        "var_A_F",
-        "var_B_A",
-        "var_B_B",
-        "var_B_C",
-        "var_B_D",
-        "var_B_E",
-        "var_B_F",
-    ]
+    assert encoder.get_feature_names() == X.columns.tolist()
     assert "var_B" not in X.columns
     assert "var_B_G" not in X.columns
     assert "var_C" in X.columns
@@ -198,20 +162,6 @@ def test_encode_top_categories():
     # note the order, it's a important
     # features are sorted in the by the name of the input column
     # and then by the number of occurrences of the category
-    assert encoder.get_feature_names() == [
-        "var_A_B",
-        "var_A_D",
-        "var_A_A",
-        "var_A_G",
-        "var_B_A",
-        "var_B_D",
-        "var_B_B",
-        "var_B_G",
-        "var_C_C",
-        "var_C_D",
-        "var_C_B",
-        "var_C_G",
-    ]
     assert encoder.get_feature_names() == X.columns.tolist()
     assert "var_B" not in X.columns
     assert "var_B_F" not in X.columns
@@ -269,14 +219,6 @@ def test_encode_numerical_variables(df_enc_numeric):
     assert encoder.encoder_dict_ == {"var_A": [1, 2, 3], "var_B": [1, 2, 3]}
     # test transform output
     pd.testing.assert_frame_equal(X, transf)
-    assert encoder.get_feature_names() == [
-        "var_A_1",
-        "var_A_2",
-        "var_A_3",
-        "var_B_1",
-        "var_B_2",
-        "var_B_3",
-    ]
     assert encoder.get_feature_names() == X.columns.tolist()
 
 
@@ -311,14 +253,6 @@ def test_variables_cast_as_category(df_enc_numeric):
     assert encoder.encoder_dict_ == {"var_A": [1, 2, 3], "var_B": [1, 2, 3]}
     # test transform output
     pd.testing.assert_frame_equal(X, transf)
-    assert encoder.get_feature_names() == [
-        "var_A_1",
-        "var_A_2",
-        "var_A_3",
-        "var_B_1",
-        "var_B_2",
-        "var_B_3",
-    ]
     assert encoder.get_feature_names() == X.columns.tolist()
 
 
@@ -368,15 +302,7 @@ def test_encode_into_k_dummy_plus_drop_binary(df_enc_binary):
     }
     # test transform output
[…]
