### 2021-08-10 d452ef2
Update base transformer with new method
+ update tests as well
