+++ tests/test_discretisation/test_equal_width_discretiser.py
@@ -29,6 +29,7 @@ def test_automatically_find_variables_and_return_as_numeric(df_normal_dist):
     # transform params
     assert (transformer.binner_dict_["var"] == bins).all()
     assert all(x for x in X["var"].unique() if x not in X_t)
+    assert transformer.get_feature_names() == X.columns.tolist()
     # in equal width discretisation, intervals get different number of values
     assert all(x for x in X["var"].value_counts() if x not in val_counts)
 
