### 2021-08-10 0b28de1
Update base transformer with new method
+ update tests as well
