+++ feature_engine/preprocessing/__init__.py
@@ -1,5 +1,5 @@
 """
-The module sanity_check includes classes to check that train and tests are similar.
+The module preprocessing includes classes to check that train and tests are similar.
 """
 
 from .similar_columns import SimilarColumns
