### 2021-08-16 9455af5
Replaced Any with Union[np.nan, int] for fill_value
