### 2021-08-16 c1b6337
Added check for fill_value Type
