### 2021-08-21 #3592 @osorensen
@Jose-Augusto-C-M @bobturneruk @JYProjs @papachristoumarios @ftarlaci @TahiriNadia would any of you be interested in reviewing this submission to the Journal of Open Source Software?
