### 2021-08-21 #3592 @osorensen
Thanks a lot, @Jose-Augusto-C-M! We have reviewer guidelines here https://joss.readthedocs.io/en/latest/reviewer_guidelines.html, which describe how to do this. In addition, the review will be in a GitHub issue, where you have a checklist to follow (see an example here https://github.com/openjournals/joss-reviews/issues/3459). In addition to that, you are welcome to contact me any time in the review process.
