### 2021-08-21 #3592 @papachristoumarios
@osorensen I can review this article. It's my first time reviewing for JOSS and may take some time. I will also read the aforementioned review guidelines.
