### 2021-08-21 #3592 @bobturneruk
Also a first timer, but the checklist looks to make the process objective and straightforward. I've used similar checklists. I'm happy to review.
