### 2021-08-21 #3592 @TahiriNadia
> @Jose-Augusto-C-M @bobturneruk @JYProjs @papachristoumarios @ftarlaci @TahiriNadia would any of you be interested in reviewing this submission to the Journal of Open Source Software?

First timer at JOSS, but the checklist makes the process seem objective and straightforward. I would be happy to review.
