### 2021-08-25 feature-engine/feature_engine#302 @papachristoumarios
Imputation and data generation methods request
Methods for _data imputation_ and _data generation_.

**Describe the solution you'd like**

1. It would be a good idea to have a data imputation method like the Amelia package in R (https://cran.r-project.org/web/packages/Amelia/index.html) together with existing imputation methods.
2. Regarding artificial data generation methods, I believe that feature-engine will benefit from copula methods such as https://sdv.dev/Copulas/ (and its related works).
