### 2021-08-25 #3642 @papachristoumarios
@osorensen My review is complete via the checklist above. Briefly, I liked the paper and the functionality it provides on a data scientist's pipeline. The software is widely used, well-structured and passes all the quality criteria outlined by the JOSS guidelines.

Finally, as a request for _future work_ I have opened an issue (https://github.com/feature-engine/feature_engine/issues/302) requesting a data imputation method and data generation methods.
