### 2021-08-30 feature-engine/feature_engine#302 @T-rect
Hi!

I installed feature_engine using pip install feature_engine. Then I created the following script and run it, but in jupyter notebook the following error appears. I tried running the script in PyScript and it runs without the error. I'm running Python 3.9.6 on Windows 10. Please help fix this problem. Thanks a lot!

import pandas as pd
import numpy as np
from feature_engine.imputation import EndTailImputer
imputer = EndTailImputer(imputation_method='gaussian', tail='right')
imputer.fit(data) #data is a Dataframe
test_t = imputer.transform(data)
test_t

<ipython-input-23-555bc2d18e70> in <module>()
      1 import pandas as pd
      2 import numpy as np
----> 3 from feature_engine.imputation import EndTailImputer
      4
      5 imputer = EndTailImputer(imputation_method='gaussian', tail='right')

ModuleNotFoundError: No module named 'feature_engine'
