### 2021-08-30 feature-engine/feature_engine#302 @solegalli
Hi @T-rect could you please delete your comment from this issue and create a new issue with your question? then i can answer there separately.

Thank you
