+++ paper/paper.bib
@@ -1,10 +1,17 @@
 @article{feature_engine_zenodo,
-  author      = {Galli, Soledad and others},
-  title       = {Feature-engine},
-  month       = Aug,
-  year        = 2021,
-  doi         = {10.5281/zenodo.5167308 },
-  url         = {https://zenodo.org/record/5167308}
+  doi         = {10.5281/zenodo.5167308},
+  url         = {https://zenodo.org/record/5167308},
+  author      = {Galli, Soledad and Gromiec, Michał and Samiullah, Christopher and
+  Kumar, Ashok and Kothareddy, Karthik and Manani, Kishan and Trejo Marrufo, Miguel and
+  Ben Driss, Sana and Krishnamurthy, Surya and Pradumna, Suryawanshi and Tan, Andrew
+  and Cabral, Arthur and Levitskiy, Gleb and Navarro Vidal, Indy and Galli, Nicolas and
+  Okroshiashvili, Nodar and Suwandi, Richard and Shah, Tejash and Agherdien, Waheeb and
+  El Amraoui, Sohayb and Patiño, Hector},
+  title       = {feature-engine/feature_engine: v1.1.1},
+  month       = {Aug},
+  year        = {2021},
+  publisher   = {Zenodo},
+  copyright   = {Open Access}
 }
 
 
