### 2021-08-30 #3592 @solegalli
>
>
> @solegalli, please also update the BibTeX entry for the Zenode archive of the package, to something like this, making sure all contributor names are included.
>
> ```
> @misc{cite_key,
>   doi = {10.5281/ZENODO.5167307},
>   url = {https://zenodo.org/record/5167307},
>   author = {Galli,  Soledad and {Michał Gromiec} and Samiullah,  Christopher and Kumar,  Ashok and Kothareddy,  Karthik and Manani,  Kishan and Marrufo,  Miguel Trejo and Driss,  Sana Ben and Krishnamurthy,  Surya and {Pradumna123} and Tan,  Andrew and Cabral,  Arthur and Levitskiy,  Gleb and {Indy Navarro Vidal} and {NicoGalli} and Okroshiashvili,  Nodar and Suwandi,  Richard Cornelius and {Tejash Shah} and {Wahe3bru} and Sohayb,  El Amraoui and {Hectorpatino} and {Piecot}},
>   title = {feature-engine/feature_engine: v1.1.1},
>   publisher = {Zenodo},
>   year = {2021},
>   copyright = {Open Access}
> }
> ```
>
> If possible, please also change the GitHub usernames appearing in this citation (Pradumna123, Wahe3bru) to the authors' real names.
>
> After you've done this, feel free to run "(at)whedon generate pdf from branch paper" as a comment in this issue, to see how the resulting paper looks.

Done! thank you for the patience.
