### 2021-08-30 feature-engine/feature_engine#302 @T-rect
Hi, ok I understand, and will do as soon as possible.
Thanks for the reply

Pada tanggal Sen, 30 Agu 2021 17.44, Soledad Galli ***@***.***>
menulis:

> Hi @T-rect <https://github.com/T-rect> could you please delete your
> comment from this issue and create a new issue with your question? then i
> can answer there separately.
>
> Thank you
>
