+++ tests/test_preprocessing/test_match_columns.py
@@ -1,16 +1,35 @@
 import numpy as np
 import pandas as pd
+import pytest
+from sklearn.exceptions import NotFittedError
 
 from feature_engine.preprocessing import MatchColumnsToTrainSet
 
+_params_fill_value = [
+    (1, [1, 1, 1, 1]),
+    (0.1, [0.1, 0.1, 0.1, 0.1]),
+    ("none", ["none", "none", "none", "none"]),
+    (np.nan, [np.nan, np.nan, np.nan, np.nan]),
+]
 
-def test_match_columns_when_more_columns_in_train_than_test(df_vartypes, df_na):
+_params_allowed = [
+    ([0, 1], "ignore", True),
+    ("nan", "hola", True),
+    ("nan", "ignore", "hallo"),
+]
 
-    # When columns are the same
+
+@pytest.mark.parametrize("fill_value, expected_studies", _params_fill_value)
+def test_columns_addition_when_more_columns_in_train_than_test(
+    fill_value, expected_studies, df_vartypes, df_na
+):
     train = df_na.copy()
     test = df_vartypes.copy()
 
-    match_columns = MatchColumnsToTrainSet(missing_values="ignore")
+    match_columns = MatchColumnsToTrainSet(
+        fill_value=fill_value,
+        missing_values="ignore",
+    )
     match_columns.fit(train)
 
     transformed_df = match_columns.transform(test)
@@ -19,52 +38,90 @@ def test_match_columns_when_more_columns_in_train_than_test(df_vartypes, df_na):
         {
             "Name": ["tom", "nick", "krish", "jack"],
             "City": ["London", "Manchester", "Liverpool", "Bristol"],
-            "Studies": [np.nan, np.nan, np.nan, np.nan],
+            "Studies": expected_studies,
             "Age": [20, 21, 19, 18],
             "Marks": [0.9, 0.8, 0.7, 0.6],
             "dob": pd.date_range("2020-02-24", periods=4, freq="T"),
         }
     )
 
-    pd.testing.assert_frame_equal(expected_result, transformed_df)
-    assert list(match_columns.variables_) == list(train.columns)
-    assert match_columns.fill_value is np.nan
+    # test init params
+    if fill_value is np.nan:
+        assert match_columns.fill_value is np.nan
+    else:
+        assert match_columns.fill_value == fill_value
     assert match_columns.verbose is True
     assert match_columns.missing_values == "ignore"
+    # test fit attrs
+    assert list(match_columns.variables_) == list(train.columns)
+    assert match_columns.n_features_in_ == 6
+    # test transform output
+    pd.testing.assert_frame_equal(expected_result, transformed_df)
 
 
-def test_match_columns_when_more_columns_in_test_than_train(df_vartypes, df_na):
-
-    # When columns are the same
-    train = df_vartypes
-    test = df_na
+def test_drop_columns_when_more_columns_in_test_than_train(df_vartypes, df_na):
+    train = df_vartypes.copy()
+    test = df_na.copy()
 
     match_columns = MatchColumnsToTrainSet(missing_values="ignore")
     match_columns.fit(train)
 
     transformed_df = match_columns.transform(test)
 
-    expected_result = pd.DataFrame(
-        {
-            "Name": ["tom", "nick", "krish", np.nan, "peter", np.nan, "fred", "sam"],
-            "City": [
-                "London",
-                "Manchester",
-                np.nan,
-                np.nan,
-                "London",
-                "London",
-                "Bristol",
-                "Manchester",
-            ],
-            "Age": [20, 21, 19, np.nan, 23, 40, 41, 37],
-            "Marks": [0.9, 0.8, 0.7, np.nan, 0.3, np.nan, 0.8, 0.6],
[…]
