+++ feature_engine/preprocessing/match_columns.py
@@ -5,12 +5,8 @@ import pandas as pd
 from sklearn.base import BaseEstimator, TransformerMixin
 from sklearn.utils.validation import check_is_fitted
 
+from feature_engine.dataframe_checks import _check_contains_na, _is_dataframe
 from feature_engine.validation import _return_tags
-
-from feature_engine.dataframe_checks import (
-    _check_contains_na,
-    _is_dataframe,
-)
 
 
 class MatchColumnsToTrainSet(BaseEstimator, TransformerMixin):
