+++ feature_engine/preprocessing/match_columns.py
@@ -11,44 +11,76 @@ from feature_engine.validation import _return_tags
 
 class MatchColumnsToTrainSet(BaseEstimator, TransformerMixin):
     """
-    MatchColumnsToTrainSet() ensure that similar columns are in test and train dataset.
+    MatchColumnsToTrainSet() ensures that the same variables observed in the train set
+    are present in the test set. If the dataset to transform contains variables that
+    were not present in the train set, they are dropped. If the dataset to transform
+    lacks variables that were present in the train set, these variables are added to
+    the dataframe with a value determined by the user (np.nan by default).
+
+    .. code-block:: python
+
+        train = pd.DataFrame({
+            "Name": ["tom", "nick", "krish", "jack"],
+            "City": ["London", "Manchester", "Liverpool", "Bristol"],
+            "Age": [20, 21, 19, 18],
+            "Marks": [0.9, 0.8, 0.7, 0.6],
+        })
+
+        test = pd.DataFrame({
+            "Name": ["tom", "sam", "nick"],
+            "Age": [20, 22, 23],
+            "Marks": [0.9, 0.7, 0.6],
+            "Hobbies": ["tennis", "rugby", "football"]
+        })
+
+        match_columns = MatchColumnsToTrainSet()
+
+        match_columns.fit(train)
+
+        df_transformed = match_columns.transform(test)
+
+        df_transformed
+
+            Name    City  Age  Marks
+        0    tom  np.nan   20    0.9
+        1    sam  np.nan   22    0.7
+        2   nick  np.nan   23    0.6
+
+
+    The order of the variables in the transformed dataset is also adjusted to match
+    that observed in the train set.
 
     Parameters
     ----------
-    fill_value: Union[int, float], default=np.NaN
-        The value that will be used to replace missing values
+    fill_value: integer, float or string. Default=np.nan
+        The values for the variables that will be added to the transformed dataset.
 
-    missing_values: string, default="raise"
-        Can take "raise", "ignore". If errors should be
-        raised in case of missing value
-
-        - raise : raise errors if there is missing value.
-        - ignore : doesn't raise errors if there is missing value.
+    missing_values: string, default='ignore'
+        Indicates if missing values should be ignored or raised. If 'ignore', the
+        transformer will ignore missing data when transforming the data. If 'raise' the
+        transformer will return an error if the training or the datasets to transform
+        contain missing values.
 
     verbose: bool, default=True
-        If the output should be verbose.
+        If True, the transformer will print out the names of the variables that are
+        added and / or removed from the dataset.
 
     Attributes
     ----------
-    fill_value:
-        The value that will be used to replace missing values
+    variables_:
+        The variables present in the train set, in the order observed during fit.
 
-    missing_values:
-        - raise : raise errors if there is missing value.
-        - ignore : doesn't raise errors if there is missing value.
-
-    verbose:
-        If the output should be verbose.
+    n_features_in_:
+        The number of features in the train set used in fit.
 
     Methods
     -------
     fit:
-        Find columns that are in the train set.
+        Identify the variable names in the train set.
     transform:
-        Add or delete columns to match train set that was called in fit.
+        Add or delete variables to match those observed in the train set.
     fit_transform:
         Fit to the data. Then transform it.
-
     """
 
     def __init__(
@@ -81,7 +113,7 @@ class MatchColumnsToTrainSet(BaseEstimator, TransformerMixin):
         self.verbose = verbose
 
     def fit(self, X: pd.DataFrame, y: pd.Series = None):
-        """Fit columns schema
+        """Learns and stores the names of the variables in the training dataset.
 
         Parameters
         ----------
@@ -92,6 +124,14 @@ class MatchColumnsToTrainSet(BaseEstimator, TransformerMixin):
         y: None
[…]
