### 2021-08-31 5a92850
fixes bug in ohe drop_binary
