+++ tests/test_encoding/test_onehot_encoder.py
@@ -40,7 +40,13 @@ def test_encode_categories_in_k_binary_plus_select_vars_automatically(df_enc_big
     }
 
     assert encoder.variables_ == ["var_A", "var_B", "var_C"]
+    assert encoder.variables_binary_ == []
     assert encoder.n_features_in_ == 3
+    assert encoder.encoder_dict_ == {
+        "var_A": ["A", "B", "C", "D", "E", "F", "G"],
+        "var_B": ["A", "B", "C", "D", "E", "F", "G"],
+        "var_C": ["A", "B", "C", "D", "E", "F", "G"],
+    }
     # test transform output
     assert X.sum().to_dict() == transf
     assert "var_A" not in X.columns
@@ -75,7 +81,12 @@ def test_encode_categories_in_k_minus_1_binary_plus_list_of_variables(df_enc_big
     }
 
     assert encoder.variables_ == ["var_A", "var_B"]
+    assert encoder.variables_binary_ == []
     assert encoder.n_features_in_ == 3
+    assert encoder.encoder_dict_ == {
+        "var_A": ["A", "B", "C", "D", "E", "F"],
+        "var_B": ["A", "B", "C", "D", "E", "F"],
+    }
     # test transform output
     for col in transf.keys():
         assert X[col].sum() == transf[col]
@@ -84,30 +95,65 @@ def test_encode_categories_in_k_minus_1_binary_plus_list_of_variables(df_enc_big
     assert "var_C" in X.columns
 
 
-def test_encode_top_categories(df_enc_big):
+def test_encode_top_categories():
     # test case 3: encode only the most popular categories
+
+    df = pd.DataFrame(
+        {
+            "var_A": ["A"] * 5
+            + ["B"] * 11
+            + ["C"] * 4
+            + ["D"] * 9
+            + ["E"] * 2
+            + ["F"] * 2
+            + ["G"] * 7,
+            "var_B": ["A"] * 11
+            + ["B"] * 7
+            + ["C"] * 4
+            + ["D"] * 9
+            + ["E"] * 2
+            + ["F"] * 2
+            + ["G"] * 5,
+            "var_C": ["A"] * 4
+            + ["B"] * 5
+            + ["C"] * 11
+            + ["D"] * 9
+            + ["E"] * 2
+            + ["F"] * 2
+            + ["G"] * 7,
+        }
+    )
+
     encoder = OneHotEncoder(top_categories=4, variables=None, drop_last=False)
-    X = encoder.fit_transform(df_enc_big)
+    X = encoder.fit_transform(df)
 
     # test init params
     assert encoder.top_categories == 4
     # test fit attr
     transf = {
-        "var_A_D": 10,
-        "var_A_B": 10,
-        "var_A_A": 6,
-        "var_A_G": 6,
-        "var_B_A": 10,
-        "var_B_D": 10,
-        "var_B_G": 6,
-        "var_B_B": 6,
-        "var_C_D": 10,
-        "var_C_C": 10,
-        "var_C_G": 6,
-        "var_C_B": 6,
[…]
