### 2021-08-31 7169e08
fixes bug in ohe drop_binary (#304)
* fixes bug in ohe drop_binary

* augments version for release
