### 2021-08-31 f732631
fixes bug in ohe drop_binary (#304)
* fixes bug in ohe drop_binary

* augments version for release
