### 2021-08-31 feature-engine/feature_engine#302 @solegalli
Thank you @papachristoumarios

At the moment, we don't have methods for artificial data generation in the roadmap for Feature-engine.

Re: the imputation method, after a quick look, am I alright to say that these are methods for time series data?

At the moment, Feature-engine takes care of tabular (cross-sectional data). We do have a module on time series in the roadmap, it is not the top priority, but it is in the horizon for consideration.

Would appreciate a bit more info on the techniques suggested, a small explanation for the lay audience, without having to read the docs would help us understand if it is in or out of scope and where in the roadmap this should be placed.

Thank you
