### 2021-09-04 #3642 @osorensen
Thanks for letting us know @Jose-Augusto-C-M. Does nothing happen if you try to click the check boxes? Could you please give it a try with a different browser? If that doesn't work I'll ask for help.
