### 2021-09-06 #3642 @bobturneruk
> :wave: @bobturneruk, please update us on how your review is going (this is an automated reminder).

Thanks @whedon. Started looking at the COI statement on Friday.
