### 2021-09-06 #3642 @Jose-Augusto-C-M
Thanks, I´ve changed the browser for Mozilla and now I´m capable of checking the boxes. I was using google chrome before.
