### 2021-09-06 #3642 @osorensen
> Thanks, I´ve changed the browser for Mozilla and now I´m capable of checking the boxes. I was using google chrome before.

Sorry about that, but glad to hear that it works for you now.
