### 2021-09-06 #3642 @Jose-Augusto-C-M
I´m surprised that the review process was so fast since i´ve started to do. I´ve tried all the codes from the examples folders and all of then work, The porpoise of the library is well defined and well written.

Thanks for considerate me for the task, I´m ready for another.
