### 2021-09-07 56238f7
Added drop and addition for multiple columns
