+++ tests/test_preprocessing/test_match_columns.py
@@ -6,10 +6,10 @@ from sklearn.exceptions import NotFittedError
 from feature_engine.preprocessing import MatchColumnsToTrainSet
 
 _params_fill_value = [
-    (1, [1, 1, 1, 1]),
-    (0.1, [0.1, 0.1, 0.1, 0.1]),
-    ("none", ["none", "none", "none", "none"]),
-    (np.nan, [np.nan, np.nan, np.nan, np.nan]),
+    (1, [1, 1, 1, 1], [1, 1, 1, 1]),
+    (0.1, [0.1, 0.1, 0.1, 0.1], [0.1, 0.1, 0.1, 0.1]),
+    ("none", ["none", "none", "none", "none"], ["none", "none", "none", "none"]),
+    (np.nan, [np.nan, np.nan, np.nan, np.nan], [np.nan, np.nan, np.nan, np.nan]),
 ]
 
 _params_allowed = [
@@ -19,12 +19,13 @@ _params_allowed = [
 ]
 
 
-@pytest.mark.parametrize("fill_value, expected_studies", _params_fill_value)
+@pytest.mark.parametrize("fill_value, expected_studies, expected_cities", _params_fill_value)
 def test_columns_addition_when_more_columns_in_train_than_test(
-    fill_value, expected_studies, df_vartypes, df_na
+    fill_value, expected_studies, expected_cities, df_vartypes, df_na
 ):
     train = df_na.copy()
     test = df_vartypes.copy()
+    test = test.drop("City", axis=1) # to add more than one column
 
     match_columns = MatchColumnsToTrainSet(
         fill_value=fill_value,
@@ -37,7 +38,7 @@ def test_columns_addition_when_more_columns_in_train_than_test(
     expected_result = pd.DataFrame(
         {
             "Name": ["tom", "nick", "krish", "jack"],
-            "City": ["London", "Manchester", "Liverpool", "Bristol"],
+            "City": expected_cities,
             "Studies": expected_studies,
             "Age": [20, 21, 19, 18],
             "Marks": [0.9, 0.8, 0.7, 0.6],
@@ -58,9 +59,9 @@ def test_columns_addition_when_more_columns_in_train_than_test(
     # test transform output
     pd.testing.assert_frame_equal(expected_result, transformed_df)
 
-
 def test_drop_columns_when_more_columns_in_test_than_train(df_vartypes, df_na):
     train = df_vartypes.copy()
+    train = train.drop("City", axis=1) # to remove more than one column
     test = df_na.copy()
 
     match_columns = MatchColumnsToTrainSet(missing_values="ignore")
@@ -68,7 +69,7 @@ def test_drop_columns_when_more_columns_in_test_than_train(df_vartypes, df_na):
 
     transformed_df = match_columns.transform(test)
 
-    expected_result = test.drop(columns=["Studies"])
+    expected_result = test.drop(columns=["Studies", "City"])
 
     # test init params
     assert match_columns.fill_value is np.nan
@@ -76,7 +77,7 @@ def test_drop_columns_when_more_columns_in_test_than_train(df_vartypes, df_na):
     assert match_columns.missing_values == "ignore"
     # test fit attrs
     assert list(match_columns.input_features_) == list(train.columns)
-    assert match_columns.n_features_in_ == 5
+    assert match_columns.n_features_in_ == 4
     # test transform output
     pd.testing.assert_frame_equal(expected_result, transformed_df)
 
