### 2021-09-07 be9ef20
Added test for simultaneous addition and removal of columns
