### 2021-09-08 7344cda
feat: High level structure of the code
