+++ feature_engine/selection/drop_high_PSI_features.py
@@ -1,0 +1,304 @@
+from typing import List, Union
+
+import pandas as pd
+
+from feature_engine.dataframe_checks import _check_contains_na, _is_dataframe
+from feature_engine.selection.base_selector import BaseSelector
+from feature_engine.validation import _return_tags
+from feature_engine.variable_manipulation import (
+    _check_input_parameter_variables,
+    _find_all_variables,
+)
+
+Variables = Union[None, int, str, List[Union[str, int]]]
+
+
+class DropHighPSIFeatures(BaseSelector):
+    """
+    
+
+    Parameters
+    ----------
+    
+
+    Attributes
+    ----------
+    features_to_drop_:
+        List with constant and quasi-constant features.
+
+    variables_:
+        The variables to consider for the feature selection.
+
+    n_features_in_:
+        The number of features in the train set used in fit.
+
+    Methods
+    -------
+    fit:
+        Find constant and quasi-constant features.
+    transform:
+        Remove constant and quasi-constant features.
+    fit_transform:
+        Fit to the data. Then transform it.
+
+    Notes
+    -----
+    This transformer is a similar concept to the VarianceThreshold from Scikit-learn,
+    but it evaluates number of unique values instead of variance
+
+    See Also
+    --------
+    sklearn.feature_selection.VarianceThreshold
+    """
+
+    def __init__(
+        self, 
+        variables: Variables = None, 
+        date_column: string = None,
+        date_cut_off: object = None,
+        threshold: float = 0.2, 
+        bucketer: object = LinearBucketer(n_bins=10),
+        missing_values: str = "raise"
+    ):
+
+        if not isinstance(threshold, (float, int)) or threshold < 0:
+            raise ValueError("threshold must be a float larger than 0")
+
+        if missing_values not in ["raise", "ignore", "include"]:
+            raise ValueError(
+                "missing_values takes only values 'raise', 'ignore' or " "'include'."
+            )
+        self.date_column = date_column
+        self.date_cut_off = date_cut_off
+        self.bucketer = bucketer
+        self.threshold = threshold
+        self.variables = _check_input_parameter_variables(variables)
+        self.missing_values = missing_values
+
+    def fit(self, X: pd.DataFrame, y: pd.Series = None):
+        """
+        Find constant and quasi-constant features.
+
+        Parameters
+        ----------
+        X: pandas dataframe of shape = [n_samples, n_features]
+            The input dataframe.
+        y: None
+            y is not needed for this transformer. You can pass y or None.
+
+        Returns
+        -------
+        self
+        """
+
+        # check input dataframe
+        X = _is_dataframe(X)
+
+        # find all variables or check those entered are present in the dataframe
+        self.variables_ = _find_all_variables(X, self.variables)
+
+        if self.missing_values == "raise":
+            # check if dataset contains na
+            _check_contains_na(X, self.variables_)
+
+        if self.missing_values == "include":
+            X[self.variables_] = X[self.variables_].fillna("missing_values")
+
+        # Split the dataframe into a reference and a comparison
+        reference_df, comparison_df = self._split_dataframe(X, self.date_column, self.date_cut_off)
+        # Compute the PSI
+        self.psi = self.compute_PSI(reference_df, comparison_df, self.bucketer)
+        
+        # Select features below the threshold
+        self.features_to_drop_ =self.psi[self.psi.value >= self.threshold].index.to_list()
+
+        return self
+
+    # Ugly work around to import the docstring for Sphinx, otherwise not necessary
+    def transform(self, X: pd.DataFrame) -> pd.DataFrame:
+        X = super().transform(X)
+
+        return X
+
[…]
