### 2021-09-08 9dabd1b
move psi code to the right place, simplify psi code and set attribute min value for bin
