### 2021-09-09 45a35d2
Improved unit test when verbose is True
