+++ tests/test_preprocessing/test_match_columns.py
@@ -8,8 +8,10 @@ from feature_engine.preprocessing import MatchColumnsToTrainSet
 _params_fill_value = [
     (1, [1, 1, 1, 1], [1, 1, 1, 1]),
     (0.1, [0.1, 0.1, 0.1, 0.1], [0.1, 0.1, 0.1, 0.1]),
-    ("none", ["none", "none", "none", "none"], ["none", "none", "none", "none"]),
-    (np.nan, [np.nan, np.nan, np.nan, np.nan], [np.nan, np.nan, np.nan, np.nan]),
+    ("none", ["none", "none", "none",
+              "none"], ["none", "none", "none", "none"]),
+    (np.nan, [np.nan, np.nan, np.nan,
+              np.nan], [np.nan, np.nan, np.nan, np.nan]),
 ]
 
 _params_allowed = [
@@ -19,17 +21,15 @@ _params_allowed = [
 ]
 
 
-@pytest.mark.parametrize(
-    "fill_value, expected_studies, expected_age",
-    _params_fill_value
-)
-def test_drop_and_add_columns(
-    fill_value, expected_studies, expected_age, df_vartypes, df_na
-):
+@pytest.mark.parametrize("fill_value, expected_studies, expected_age",
+                         _params_fill_value)
+def test_drop_and_add_columns(fill_value, expected_studies, expected_age,
+                              df_vartypes, df_na):
     train = df_na.copy()
     test = df_vartypes.copy()
     test = test.drop("Age", axis=1)  # to add more than one column
-    for new_col in ["test1", "test2"]:  # adding columns to test if they are removed
+    for new_col in ["test1",
+                    "test2"]:  # adding columns to test if they are removed
         test.loc[:, new_col] = new_col
 
     match_columns = MatchColumnsToTrainSet(
@@ -40,16 +40,17 @@ def test_drop_and_add_columns(
 
     transformed_df = match_columns.transform(test)
 
-    expected_result = pd.DataFrame(
-        {
-            "Name": ["tom", "nick", "krish", "jack"],
-            "City": ["London", "Manchester", "Liverpool", "Bristol"],
-            "Studies": expected_studies,
-            "Age": expected_age,
-            "Marks": [0.9, 0.8, 0.7, 0.6],
-            "dob": pd.date_range("2020-02-24", periods=4, freq="T"),
-        }
-    )
+    expected_result = pd.DataFrame({
+        "Name": ["tom", "nick", "krish", "jack"],
+        "City": ["London", "Manchester", "Liverpool", "Bristol"],
+        "Studies":
+        expected_studies,
+        "Age":
+        expected_age,
+        "Marks": [0.9, 0.8, 0.7, 0.6],
+        "dob":
+        pd.date_range("2020-02-24", periods=4, freq="T"),
+    })
 
     # test init params
     if fill_value is np.nan:
@@ -65,13 +66,10 @@ def test_drop_and_add_columns(
     pd.testing.assert_frame_equal(expected_result, transformed_df)
 
 
-@pytest.mark.parametrize(
-    "fill_value, expected_studies, expected_age",
-    _params_fill_value
-)
+@pytest.mark.parametrize("fill_value, expected_studies, expected_age",
+                         _params_fill_value)
 def test_columns_addition_when_more_columns_in_train_than_test(
-    fill_value, expected_studies, expected_age, df_vartypes, df_na
-):
+        fill_value, expected_studies, expected_age, df_vartypes, df_na):
     train = df_na.copy()
     test = df_vartypes.copy()
     test = test.drop("Age", axis=1)  # to add more than one column
@@ -84,16 +82,17 @@ def test_columns_addition_when_more_columns_in_train_than_test(
 
[…]
