### 2021-09-09 e14095c
Improved unit test when verbose is True
