### 2021-09-09 b7eb3c9
Added documentation for MatchColumnsToTrainSet
