+++ docs/preprocessing/MatchColumnsToTrainSet.rst
@@ -1,5 +1,5 @@
 MatchColumnsToTrainSet
-====================
+======================
 
 API Reference
 -------------
