+++ docs/preprocessing/index.rst
@@ -1,0 +1,16 @@
+.. -*- mode: rst -*-
+
+Preprocessing
+=============
+
+Feature-engine's preprocessing transformers are used to apply general preprocessing steps.
+
+.. figure::  ../images/selectionSummary.png
+   :align:   center
+
+   Summary of Feature-engine's preprocessing main characteristics
+
+.. toctree::
+   :maxdepth: 2
+
+   MatchColumnsToTrainSet
