+++ feature_engine/selection/drop_high_PSI_features.py
@@ -1,16 +1,140 @@
 from typing import List, Union
 
 import pandas as pd
+import numpy as np
 
 from feature_engine.dataframe_checks import _check_contains_na, _is_dataframe
 from feature_engine.selection.base_selector import BaseSelector
+from feature_engine.discretisation.equal_frequency import EqualFrequencyDiscretiser
+from feature_engine.discretisation.equal_width import EqualWidthDiscretiser
 from feature_engine.validation import _return_tags
 from feature_engine.variable_manipulation import (
     _check_input_parameter_variables,
     _find_all_variables,
+    _find_or_check_numerical_variables,
 )
 
 Variables = Union[None, int, str, List[Union[str, int]]]
+
+class dvl():
+
+    def __init__(self, compare, variables: Variables = None, missing_values: str = "raise",
+    threshold: int = 0.25,n_bins = 10, method = 'equal_frequency'):
+
+        if not isinstance(threshold, (float, int)) or threshold < 0:
+            raise ValueError("threshold must be a float larger than 0")
+
+        if missing_values not in ["raise", "ignore", "include"]:
+            raise ValueError(
+                "missing_values takes only values 'raise', 'ignore' or " "'include'."
+            )
+        
+        self._min_value = 0.0001
+        self._compare = compare
+        self.threshold = threshold
+        self.variables = _check_input_parameter_variables(variables)
+        self.missing_values = missing_values
+
+        if method.lower() in ["equalwidth", 'equal_width', "equal width"]:
+            self.bucketer = EqualWidthDiscretiser(bins=n_bins)
+        elif method.lower() in  ["equalfrequency", 'equal_frequency', "equal frequency"]:
+            self.bucketer = EqualFrequencyDiscretiser(q=n_bins)
+        else:
+            raise ValueError("Incorrect name for the method, should be either equal_width or equal_frequency")
+        
+    def fit(self, X: pd.DataFrame, y: pd.Series = None):
+        # check input dataframe
+        X = _is_dataframe(X)
+
+        # find all variables or check those entered are present in the dataframe
+        self.variables_ =  _find_or_check_numerical_variables(X, self.variables)
+
+        if self.missing_values == "raise":
+            # check if dataset contains na
+            _check_contains_na(X, self.variables_)
+
+        if self.missing_values == "include":
+            X[self.variables_] = X[self.variables_].fillna("missing_values")
+
+        # Split the dataframe into a reference and a comparison
+        reference_df, comparison_df = self._split_dataframe(X, self._compare)
+        # Compute the PSI
+        self.psi = self._compute_PSI(reference_df, comparison_df, self.bucketer)
+        
+        # Select features below the threshold
+        self.features_to_drop_ = self.psi[self.psi.value >= self.threshold].index.to_list()
+
+        return self
+
+    def _compute_PSI(self,df_ref, df_comp, bucketer):
+
+        ref = bucketer.fit_transform(df_ref)
+        comp = bucketer.transform(df_comp)
+
+        results = {}
+        
+        print(self.variables_)
+        for feature in self.variables_:
+            results[feature] = [self._compute_feature_psi(
+                ref[[feature]].value_counts(), 
+                comp[[feature]].value_counts(), bucketer)]
+
+        results_df = pd.DataFrame.from_dict(results).T
+        results_df.columns = ['value']
+
+        return results_df
+
+
+    def _compute_feature_psi(self, series_ref, series_comp, bucketer):
+
+        binning = pd.DataFrame(series_ref).merge(pd.DataFrame(series_comp), 
+        right_index=True, left_index=True, how="outer"
+        ).fillna(0)
[…]
