+++ docs/preprocessing/MatchColumnsToTrainSet.rst
@@ -11,14 +11,17 @@ API Reference
 Example
 -------
 
-The MatchColumnsToTrainSet() ensure that columns in test dataset are similar 
-to train dataset. 
-If needed it drops columns that are in test but not in the train dataset
-and add columns that are missing 
-in test dataset (but are in train).
+MatchColumnsToTrainSet() ensures that the columns in the test set are identical to those
+in the train set.
+
+If the test set contains additional columns, they are dropped. Alternatively, if the
+test set lacks columns that were present in the train set, they will be added with a
+value determined by the user, for example np.nan.
+
 
 .. code:: python
 
+    import numpy as np
     import pandas as pd
 
     from feature_engine.preprocessing import MatchColumnsToTrainSet
@@ -26,12 +29,18 @@ in test dataset (but are in train).
 
     # Load dataset
     def load_titanic():
-            data = pd.read_csv('https://www.openml.org/data/get_csv/16826755/phpMYEkMl')
-            data = data.replace('?', np.nan)
-            data['cabin'] = data['cabin'].astype(str).str[0]
-            data['pclass'] = data['pclass'].astype('O')
-            data['embarked'].fillna('C', inplace=True)
-            return data
+        data = pd.read_csv('https://www.openml.org/data/get_csv/16826755/phpMYEkMl')
+        data = data.replace('?', np.nan)
+        data['cabin'] = data['cabin'].astype(str).str[0]
+        data['pclass'] = data['pclass'].astype('O')
+        data['age'] = data['age'].astype('float')
+        data['fare'] = data['fare'].astype('float')
+        data['embarked'].fillna('C', inplace=True)
+        data.drop(
+            labels=['name', 'ticket', 'boat', 'body', 'home.dest'],
+            axis=1, inplace=True,
+        )
+        return data
 
     # load data as pandas dataframe
     data = load_titanic()
@@ -40,38 +49,121 @@ in test dataset (but are in train).
     train = data.iloc[0:1000, :]
     test = data.iloc[1000:, :]
 
-    columns_matcher = MatchColumnsToTrainSet(missing_values="ignore")
+    # set up the transformer
+    match_cols = MatchColumnsToTrainSet(missing_values="ignore")
 
-    # We drop some columns to test the transformer
-    test_with_missing_columns = test.drop(["sex", "ticket"], axis=1)
+    # learn the variables in the train set
+    match_cols.fit(train)
 
-    test_with_missing_columns.head()
+    # the transformer stores the input variables
+    match_cols.input_features_
+
 
 .. code:: python
 
-    pclass	survived	fare	cabin	embarked
-    1000	3	1	7.7500	n	Q
-    1001	3	1	23.2500	n	Q
-    1002	3	1	23.2500	n	Q
-    1003	3	1	23.2500	n	Q
-    1004	3	1	7.7875	n	Q
+    ['pclass',
+     'survived',
+     'sex',
+     'age',
+     'sibsp',
+     'parch',
+     'fare',
+     'cabin',
+     'embarked']
+
 
 .. code:: python
 
-    columns_matcher.fit(train)
+    # Let's drop some columns in the test set for the demo
+    test_t = test.drop(["sex", "age"], axis=1)
 
-    res = columns_matcher.transform(test_with_missing_columns)
-
-    res.head()
+    test_t.head()
 
 .. code:: python
 
-    pclass	survived	sex	ticket	fare	cabin	embarked
-    1000	3	1	NaN	NaN	7.7500	n	Q
-    1001	3	1	NaN	NaN	23.2500	n	Q
-    1002	3	1	NaN	NaN	23.2500	n	Q
-    1003	3	1	NaN	NaN	23.2500	n	Q
[…]
