+++ docs/preprocessing/index.rst
@@ -9,4 +9,4 @@ and transformation procedures.
 .. toctree::
    :maxdepth: 2
 
-   MatchColumnsToTrainSet
+   MatchVariables
