### 2021-09-10 497909e
add preprocessing to tree index
