+++ docs/index.rst
@@ -242,6 +242,7 @@ The `issues <https://github.com/feature-engine/feature_engine/issues/>`_ and
    outliers/index
    creation/index
    selection/index
+   preprocessing/index
    wrappers/index
    preprocessing/index
 
