### 2021-09-10 511cb1f
add preprocessing to tree index
