### 2021-09-10 7851e43
Removed examples folder
