### 2021-09-10 15934a1
feat: clean code and add tests
