+++ feature_engine/selection/drop_high_PSI_features.py
@@ -16,33 +16,50 @@ from feature_engine.variable_manipulation import (
 
 Variables = Union[None, int, str, List[Union[str, int]]]
 
-class dvl():
+class DropHighPSIFeatures(BaseSelector):
+    """
+    DropHighPSIFeatures drops features with a Population Stability Index (PSI) above
+    a given threshold. The PSI of a feature is an indication of the shift in its 
+    distribution; a feature with high PSI might therefore be seen as instable.
+
+    In fields like Credit Risk Modelling, the elimination of features with high PSI
+    is frequent and usually required by the Regulator. 
+
+    The calculation of the PSI requires to compare two distributions. In DropHighPSIFeatures
+    two approaches are implemented though the "basis" argument.
+
+    - If it is a pandas.DataFrame, the class will compare the distributions of the X dataframe
+    (argument of the fit method) and "basis". The two dataframes must contain the same features
+    (i.e. labels). 
+
+    - If it is a dictionary, the X matrix of the fit method is split in two according to time
+    and the distribution between the two parts are compared using the PSI. The dictionary must 
+    contain the label of a column with dates and the cut-off date.
+
+    - The PSI calculations are not symmetric. The switch_basis argument allows to switch the 
+    role of the two dataframes in the PSI calculations.
+
+    The comparison of the distribution is done through binning. Two strategies are implemented:
+    equal_frequency and equal_width. These labels refer to two discretisation implementation from
+    the present package. 
+
+    References:
+    - [Statistical Properties of Population Stability Index](https://scholarworks.wmich.edu/cgi/viewcontent.cgi?article=4249&context=dissertations)
+
+    """
 
     def __init__(self, basis, variables: Variables = None, missing_values: str = "raise",
-    switch_basis=False, threshold: int = 0.25,n_bins = 10, method = 'equal_frequency'):
+    switch_basis=False, threshold: int = 0.25,n_bins = 10, method = 'equal_frequency',
+    min_pct_empty_buckets = 0.0001):
 
-        if not isinstance(threshold, (float, int)) or threshold < 0:
-            raise ValueError("threshold must be a float larger than 0")
-
-        if missing_values not in ["raise", "ignore", "include"]:
-            raise ValueError(
-                "missing_values takes only values 'raise', 'ignore' or " "'include'."
-            )
+        # Set all arguments (except self and variables as attributes)
+        [setattr(self,name,value) for name,value in vars().items() if name not in ("self", 'variables')]
         
-        self._min_value = 0.0001
-        self._basis = basis
-        self.switch_basis = switch_basis
-        self.threshold = threshold
         self.variables = _check_input_parameter_variables(variables)
-        self.missing_values = missing_values
 
-        if method.lower() in ["equalwidth", 'equal_width', "equal width"]:
-            self.bucketer = EqualWidthDiscretiser(bins=n_bins)
-        elif method.lower() in  ["equalfrequency", 'equal_frequency', "equal frequency"]:
-            self.bucketer = EqualFrequencyDiscretiser(q=n_bins)
-        else:
-            raise ValueError("Incorrect name for the method, should be either equal_width or equal_frequency")
-        
+        self._check_init_values(basis, variables, missing_values, switch_basis, 
+        threshold, n_bins, method, min_pct_empty_buckets)
+
     def fit(self, X: pd.DataFrame, y: pd.Series = None):
         # check input dataframe
         X = _is_dataframe(X)
@@ -58,7 +75,7 @@ class dvl():
             X[self.variables_] = X[self.variables_].fillna("missing_values")
 
[…]
