+++ tests/test_selection/test_drop_high_psi_features.py
@@ -1,0 +1,177 @@
+import pandas as pd
+import pytest
+from sklearn.datasets import make_classification
+from sklearn.exceptions import NotFittedError
+from datetime import date
+
+from feature_engine.selection.drop_high_PSI_features import DropHighPSIFeatures
+
+
+@pytest.fixture(scope="module")
+def df():
+    # create array with 4 correlated features and 2 independent ones
+    X, y = make_classification(
+        n_samples=1000,
+        n_features=6,
+        n_redundant=2,
+        n_clusters_per_class=1,
+        weights=[0.50],
+        class_sep=2,
+        random_state=1,
+    )
+
+    # transform array into pandas df
+    colnames = ["var_" + str(i) for i in range(6)]
+    X = pd.DataFrame(X, columns=colnames)
+
+    return X
+
+
+@pytest.fixture(scope="module")
+def df_basis():
+    # create array with 8 correlated features and 4 independent ones
+    X, y = make_classification(
+        n_samples=1000,
+        n_features=12,
+        n_redundant=4,
+        n_clusters_per_class=1,
+        weights=[0.50],
+        class_sep=2,
+        random_state=1,
+    )
+
+    # transform array into pandas df
+    colnames = ["var_" + str(i) for i in range(12)]
+    X = pd.DataFrame(X, columns=colnames)
+
+    return X
+
+
+def test_sanity_checks(df_basis):
+    """Sanity checks.
+
+    All PSI must be zero if the two dataframe involved are the same.
+    There will be no changes in the dataframe.
+    """
+    transformer = DropHighPSIFeatures(df_basis)
+    X = transformer.fit_transform(df_basis)
+
+    assert transformer.psi.value.min() == 0
+    assert transformer.psi.value.max() == 0
+    assert X.shape == df_basis.shape
+    assert transformer.psi.shape[0] == df_basis.shape[1] 
+
+def test_check_psi_values():
+    """Compare the PSI value with a reference.
+
+    As reference we take the implementation from the Probatus package.
+    ```
+    from probatus.stat_tests import AutoDist
+    
+    df = pd.DataFrame({"A": [1, 1, 1, 4]})
+    df2 = pd.DataFrame({"A": [4, 4, 4, 1]})
+
+    (AutoDist(statistical_tests=['PSI'], binning_strategies="SimpleBucketer", bin_count=2)
+    .compute(df, df2)
+    )
+    ```
+    Output; 1.098612
+    """
+    df = pd.DataFrame({"A": [1, 1, 1, 4]})
+    df2 = pd.DataFrame({"A": [4, 4, 4, 1]})
+
+    ref_value = 1.098612
+
+    test=DropHighPSIFeatures(
+        df2, n_bins=2, 
+        method='equal_width', switch_basis=False, 
+        min_pct_empty_buckets = 0.001)
+
+    test.fit(df)
+    assert abs(test.psi.value[0] - ref_value) < 0.000001
+
+
+def test_switch_basis():
+    """Test the functionality to switch the basis."""
+    import pandas as pd
+
+    df_a = pd.DataFrame({
+        "a": [1., 2, 3, 1],
+        "b": [1., 2, 3, 4],
+        "c": [1, 2, 3, 4],
+        "d": [1.7, 4.7, 6.6, 7.8]
+    })
+
+    df_b = pd.DataFrame({
+        "a": [4., 3, 5, 1],
+        "b": [11., 1, 2, 4],
+        "c": [4, 2, 2, 4],
+        "d": [4.7, 4.7, 7.6, 7.8]
+    })
+
[…]
