+++ feature_engine/selection/base_selector.py
@@ -27,7 +27,6 @@ def get_feature_importances(estimator):
 
     return list(importances)
 
-
 class BaseSelector(BaseEstimator, TransformerMixin):
     """Transformation shared by all selectors"""
 
