+++ feature_engine/selection/drop_high_PSI_features.py
@@ -46,8 +46,91 @@ class DropHighPSIFeatures(BaseSelector):
     References:
     - [Statistical Properties of Population Stability Index](https://scholarworks.wmich.edu/cgi/viewcontent.cgi?article=4249&context=dissertations)
 
-    """
+    DropHighPSIFeatures() works only with numerical variables. Categorical variables
+    will need to be encoded to numerical or will be excluded from the analysis.
 
+    Parameters
+    ----------
+
+    basis: pd.DataFrame or dictionary.
+        Information required to define the basis for the PSI calculations. It is either a
+        dataframe in the case of direct comparison or the label of the column containing the
+        date and the cut-off-date in case a single dataframe is provided. In the latter case,
+        the dataframe will be split on two parts that are non-overlapping over time. 
+
+    variables: list, default=None
+        The list of variables to evaluate. If None, the transformer will evaluate all
+        numerical variables in the dataset.
+
+    switch_basis: boolean, default=False.
+        If set to true the role of the two matrices involved in the PSI calculations
+        (basis and measurement) will be switch. This is an important option as the 
+        PSI value is not symmetric (i.e. PSI(a, b) != PSI(b, a)).
+
+    threshold: float, default = 0.25.
+        Threshold above which the distribution of a feature has changed so much that the
+        feature will be dropped. The most common values are 0.25 (large shift) and 0.10
+        (medium shift). 
+
+    method: string or callable, default='equal_frequency'
+        Type of binning used to represent the distribution of the feature. In can be either
+        "equal_width" for equally spaced bins or "equal_frequency" for bins based on quantiles.
+
+    n_nbins, int, default = 10
+        Number of bins used in the binning. For numerical feature a value of 10 is considered 
+        as appropriate. For features with lower cardinality lower values are usually used.
+
+    min_pct_empty_buckets, float, default = 0.0001
+        Value to add to empty bucket (when considering percentages). If a bin is empty the 
+        PSI value may jump to infinity. By adding a small number to empty bins, this issue 
+        is avoided. If the value added is too large, it may disturb the calculations.
+
+    missing_values: str, default=ignore
+        Takes values 'raise' and 'ignore'. Whether the missing values should be raised
+        as error or ignored when determining correlation.
+
+    Attributes
+    ----------
+    features_to_drop_:
+        Set with the correlated features that will be dropped.
+
+    correlated_feature_sets_:
+        Groups of correlated features. Each list is a group of correlated features.
+
+    variables_:
+        The variables to consider for the feature selection.
+
+    n_features_in_:
+        The number of features in the train set used in fit.
+
+    psi: 
+        Dataframe containing the PSI values for all features considered.
+
+    features_to_drop_:
+        List of features that are dropped because of PSI values above threshold.
+
+    Methods
+    -------
+    fit:
+        Find correlated features.
+    transform:
+        Remove correlated features.
+    fit_transform:
+        Fit to the data. Then transform it.
+    _compute_psi:
+        Perform binning and compute the PSI for all features
+    _compute_feature_psi:
+        Combine information and call psi calculation 
+    _psi
+        Compute the PSI value
+    _split_dataframe:
+        returns the basis and the measurement dataframes needed for PSI calculations
+    _check_init_values:
+        Perform a series of check on the attributes.
+    """
+    # TODO: Check impact of "missing_values"!
+    # TODO: Implement the check on the types of the cut-off dates and the date column that 
+    # need ot be the same.
     def __init__(self, basis, variables: Variables = None, missing_values: str = "raise",
[…]
