### 2021-09-10 1633017
bug: Move binning to feature level and drop na
