+++ feature_engine/sanity_check/__init__.py
@@ -1,9 +1,0 @@
-"""
-The module sanity_check includes classes to check that train and tests are similar.
-"""
-
-from .similar_columns import SimilarColumns
-
-__all__ = [
-    "SimilarColumns",
-]
