### 2021-09-14 7f9d7c1
Add match variables closes #132 (#308)
* Added SimilarColumns (Automatic handle of columns difference in test and train)

* Update feature_selection.py

Added typing and input check

* Update feature_selection.py

Added pandas import

* Refactoring

* Replaced BaseSelector with TransformerMixin and BaseEstimator

* Style improvement

* Improved Typing

* code review refactoring

* style

* Added drop_if_more_columns and add_if_less_columns attributes

* style

* style

* style

* style

* new way to add if less columns

* Refactoring and error handling

* style

* styler

* refactoring of test
[…]
