+++ feature_engine/preprocessing/__init__.py
@@ -1,0 +1,10 @@
+"""
+The module preprocessing includes classes and functions for general data pre-processing
+and transformation.
+"""
+
+from .match_columns import MatchVariables
+
+__all__ = [
+    "MatchVariables",
+]
