+++ feature_engine/preprocessing/match_columns.py
@@ -1,0 +1,197 @@
+from typing import List, Union
+
+import numpy as np
+import pandas as pd
+from sklearn.base import BaseEstimator, TransformerMixin
+from sklearn.utils.validation import check_is_fitted
+
+from feature_engine.dataframe_checks import _check_contains_na, _is_dataframe
+from feature_engine.validation import _return_tags
+
+
+class MatchVariables(BaseEstimator, TransformerMixin):
+    """
+    MatchVariables() ensures that the same variables observed in the train set
+    are present in the test set. If the dataset to transform contains variables that
+    were not present in the train set, they are dropped. If the dataset to transform
+    lacks variables that were present in the train set, these variables are added to
+    the dataframe with a value determined by the user (np.nan by default).
+
+    .. code-block:: python
+
+        train = pd.DataFrame({
+            "Name": ["tom", "nick", "krish", "jack"],
+            "City": ["London", "Manchester", "Liverpool", "Bristol"],
+            "Age": [20, 21, 19, 18],
+            "Marks": [0.9, 0.8, 0.7, 0.6],
+        })
+
+        test = pd.DataFrame({
+            "Name": ["tom", "sam", "nick"],
+            "Age": [20, 22, 23],
+            "Marks": [0.9, 0.7, 0.6],
+            "Hobbies": ["tennis", "rugby", "football"]
+        })
+
+        match_columns = MatchVariables()
+
+        match_columns.fit(train)
+
+        df_transformed = match_columns.transform(test)
+
+        df_transformed
+
+            Name    City  Age  Marks
+        0    tom  np.nan   20    0.9
+        1    sam  np.nan   22    0.7
+        2   nick  np.nan   23    0.6
+
+
+    The order of the variables in the transformed dataset is also adjusted to match
+    that observed in the train set.
+
+    Parameters
+    ----------
+    fill_value: integer, float or string. Default=np.nan
+        The values for the variables that will be added to the transformed dataset.
+
+    missing_values: string, default='ignore'
+        Indicates if missing values should be ignored or raised. If 'ignore', the
+        transformer will ignore missing data when transforming the data. If 'raise' the
+        transformer will return an error if the training or the datasets to transform
+        contain missing values.
+
+    verbose: bool, default=True
+        If True, the transformer will print out the names of the variables that are
+        added and / or removed from the dataset.
+
+    Attributes
+    ----------
+    input_features_:
+        The variables present in the train set, in the order observed during fit.
+
+    n_features_in_:
+        The number of features in the train set used in fit.
+
+    Methods
+    -------
+    fit:
+        Identify the variable names in the train set.
+    transform:
+        Add or delete variables to match those observed in the train set.
+    fit_transform:
+        Fit to the data. Then transform it.
+    """
+
+    def __init__(
+        self,
+        fill_value: Union[str, int, float] = np.nan,
+        missing_values: str = "raise",
+        verbose: bool = True,
+    ):
+
+        if missing_values not in ["raise", "ignore"]:
+            raise ValueError(
+                "missing_values takes only values 'raise' or 'ignore'."
+                f"Got '{missing_values} instead.")
+
+        if not isinstance(verbose, bool):
+            raise ValueError("verbose takes only booleans True and False."
+                             f"Got '{verbose} instead.")
+
+        # note: np.nan is an instance of float!!!
+        if not isinstance(fill_value, (str, int, float)):
+            raise ValueError("fill_value takes integers, floats or strings."
+                             f"Got '{fill_value} instead.")
+
+        self.fill_value = fill_value
+        self.missing_values = missing_values
+        self.verbose = verbose
+
+    def fit(self, X: pd.DataFrame, y: pd.Series = None):
[…]
