+++ tests/test_preprocessing/test_check_estimator_preprocessing.py
@@ -1,0 +1,14 @@
+import pytest
+from sklearn.utils.estimator_checks import check_estimator
+
+from feature_engine.preprocessing import MatchVariables
+
+
+@pytest.mark.parametrize(
+    "Estimator",
+    [
+        MatchVariables(),
+    ],
+)
+def test_all_transformers(Estimator):
+    return check_estimator(Estimator)
