### 2021-09-17 #3642 @osorensen
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.5281/zenodo.5167308 is OK
> -  10.25080/Majora-92bf1922-00a  is OK
> - 10.1145/507533.507538 is OK
> - 10.1145/2347736.2347755 is OK
> - 10.21105/joss.00501 is OK
>
> MISSING DOIs
>
> - 10.1093/biomet/87.4.954 may be a valid DOI for title: A New Family of Power Transformations to Improve Normality or Symmetry
> - 10.1109/dsaa.2015.7344858 may be a valid DOI for title: Deep feature synthesis: Towards automating data science endeavors
>
> INVALID DOIs
>
> - https://doi.org/10.1111/j.2517-6161.1964.tb00553.x is INVALID because of 'https://doi.org/' prefix
> - https://doi.org/10.1016/j.neucom.2018.03.067 is INVALID because of 'https://doi.org/' prefix
> ```

@solegalli, could you please correct the reference issues above?
