### 2021-09-17 #3642 @osorensen
@solegalli, please also correct the following in the manuscript:

- Line 29: change from "Feature engine" to "Feature-engine".
- Line 31: I suggest remove the comma at the end of "as well as,".
- Lines 33-34: I suggest rewriting "... the current Python’s scientific computing stack ..." to "... Python's currrent scientific computing stack ...".
- Line 45: I suggest rewriting "... data competitions ..." to "... data science competitions ...".
- Line 81: Make sure Kaggle is capitalized in the reference. You can do this by addding curly braces ({Kaggle}) to the .bib file.

Finally, could you please confirm that https://doi.org/10.5281/zenodo.5167308 is the correct link to the Zenodo archive for the submitted software?
