### 2021-09-18 db6e836
fixes spelling  and punctuation
