+++ paper/paper.bib
@@ -154,7 +154,7 @@ journal = {International Journal of Computer Science}
 
 @misc{beatingkaggle,
   author = {Ying Dong},
-  title = {Beating Kaggle the easy way},
+  title = {Beating {Kaggle} the easy way},
   year = {2015},
   publisher = {Technische Universität Darmstadt},
   journal = {Studienarbeit},
