### 2021-09-18 d48a5a7
adds missing DOIs
