+++ paper/paper.bib
@@ -123,21 +123,23 @@
 
 
 @article{yeojohnson,
- ISSN = {00063444},
- URL = {http://www.jstor.org/stable/2673623},
- abstract = {We introduce a new power transformation family which is well
- defined on the whole real line and which is appropriate for reducing skewness
- and to approximate normality. It has properties similar to those of the
- Box-Cox transformation for positive variables. The large-sample properties
- of the transformation are investigated in the contect of a single random sample},
- author = {In-Kwon Yeo and Richard A. Johnson},
- journal = {Biometrika},
- number = {4},
- pages = {954--959},
- publisher = {[Oxford University Press, Biometrika Trust]},
- title = {A New Family of Power Transformations to Improve Normality or Symmetry},
- volume = {87},
- year = {2000}
+  author = {Yeo, In‐Kwon and Johnson, Richard A.},
+  title = "{A new family of power transformations to improve normality or symmetry}",
+  journal = {Biometrika},
+  volume = {87},
+  number = {4},
+  pages = {954-959},
+  year = {2000},
+  month = {12},
+  abstract = "{We introduce a new power transformation family which is well defined on
+  the whole real line and which is appropriate for reducing skewness and to approximate
+  normality. It has properties similar to those of the Box–Cox transformation for positive
+  variables. The large‐sample properties of the transformation are investigated in the
+  context of a single random sample.}",
+  issn = {0006-3444},
+  doi = {10.1093/biomet/87.4.954},
+  url = {https://doi.org/10.1093/biomet/87.4.954},
+  eprint = {https://academic.oup.com/biomet/article-pdf/87/4/954/633221/870954.pdf},
 }
 
 
@@ -223,7 +225,8 @@ journal = {International Journal of Computer Science}
   Analytics, DSAA 2015, Paris, France, October 19-21, 2015},
   pages     = {1--10},
   year      = {2015},
-  organization={IEEE}
+  organization={IEEE},
+   doi={10.1109/DSAA.2015.7344858},
 }
 
 
