+++ paper/paper.bib
@@ -1,14 +1,14 @@
 @article{feature_engine_zenodo,
-  doi         = {10.5281/zenodo.5167308},
-  url         = {https://zenodo.org/record/5167308},
+  doi         = {10.5281/zenodo.5515531},
+  url         = {https://doi.org/10.5281/zenodo.5515531},
   author      = {Galli, Soledad and Gromiec, Michał and Samiullah, Christopher and
   Kumar, Ashok and Kothareddy, Karthik and Manani, Kishan and Trejo Marrufo, Miguel and
   Ben Driss, Sana and Krishnamurthy, Surya and Pradumna, Suryawanshi and Tan, Andrew
   and Cabral, Arthur and Levitskiy, Gleb and Navarro Vidal, Indy and Galli, Nicolas and
   Okroshiashvili, Nodar and Suwandi, Richard and Shah, Tejash and Agherdien, Waheeb and
   El Amraoui, Sohayb and Patiño, Hector},
-  title       = {feature-engine/feature_engine: v1.1.1},
-  month       = {Aug},
+  title       = {feature-engine/feature_engine: v1.1.2},
+  month       = {Sep},
   year        = {2021},
   publisher   = {Zenodo},
   copyright   = {Open Access}
