### 2021-09-18 #3642 @solegalli
> - https://doi.org/10.1111/j.2517-6161.1964.tb00553.x is INVALID because of 'https://doi.org/' prefix
> - https://doi.org/10.1016/j.neucom.2018.03.067 is INVALID because of 'https://doi.org/' prefix

The DOIs appear in the format that I have used in the .bib. In other words, they seem to be correct as per the below websites:
https://rss.onlinelibrary.wiley.com/action/showCitFormats?doi=10.1111%2Fj.2517-6161.1964.tb00553.x
https://www.sciencedirect.com/science/article/pii/S0925231218304843?via%3Dihub

I go them by visiting the websites and exporting the citation in bibtex format.

I tried removing the doi.org, but then it does not point to the right place. What shall I do?
