### 2021-09-18 9e208b8
removes doi.org from citations
