+++ paper/paper.bib
@@ -104,7 +104,7 @@
   volume = {26},
   number = {2},
   pages = {211-243},
-  doi = {https://doi.org/10.1111/j.2517-6161.1964.tb00553.x},
+  doi = {10.1111/j.2517-6161.1964.tb00553.x},
   url = {https://rss.onlinelibrary.wiley.com/doi/abs/10.1111/j.2517-6161.1964.tb00553.x},
   eprint = {https://rss.onlinelibrary.wiley.com/doi/pdf/10.1111/j.2517-6161.1964.tb00553.x},
   abstract = {Summary In the analysis of data it is often assumed that observations
@@ -238,7 +238,7 @@ journal = {International Journal of Computer Science}
   pages = {72-77},
   year = {2018},
   issn = {0925-2312},
-  doi = {https://doi.org/10.1016/j.neucom.2018.03.067},
+  doi = {10.1016/j.neucom.2018.03.067},
   url = {https://www.sciencedirect.com/science/article/pii/S0925231218304843},
   author = {Maximilian Christ and Nils Braun and Julius Neuffer and
   Andreas W. Kempa-Liehr},
