### 2021-09-20 #3642 @kthyng
Is the proper Zenodo doi 10.5281/zenodo.5515531? I clicked on the one linked above but it says there is a newer version. Also, @solegalli, please update the title and author list of the Zenodo archive metadata so they exactly match the JOSS paper.
