### 2021-09-20 #3642 @kthyng
@solegalli Your paper looks good other than the Galli et al reference which is to the software archive for this submission itself. It's not typical to cite within the submission like that, but probably ok. Right now the reference is weird though, I think because of the metadata issue I listed above. If you want to keep citing the software archive itself (which is indeed linked to separately in the paper), you'll want to update your .bib file accordingly after the metadata in Zenodo is updated.
