### 2021-09-21 #3642 @solegalli
Hi @kthyng

Thanks for the input.

I saw that these developers cite zenodo for their package [yellobricks in JOSS](https://www.theoj.org/joss-papers/joss.01075/10.21105.joss.01075.pdf). I wasn't sure what the best practice / common practice was.

I went ahead and removed the citation. The link is meant to be updated all the time as new releases are made, so I think it makes sense.

Thanks again!
