### 2021-09-21 #3642 @solegalli
v1.1.2 is the last one / current one.

When I submitted the article it was v1.1.1 but I had to fix a bug that we accidentally introduced in this version :_( .Thus the increment, it is really minor.

So to summarize, latest version v1.1.2 and the zenodo link is https://zenodo.org/record/5515531

Cheers
