### 2021-09-21 #3642 @kthyng
Ok great, thanks. One more thing: please update the title and author list of the Zenodo archive metadata so they exactly match the JOSS paper.
