### 2021-09-21 #3642 @solegalli
>
>
> Ok great, thanks. One more thing: please update the title and author list of the Zenodo archive metadata so they exactly match the JOSS paper.

I am not sure what I should do. Should I delete all the other authors from zenodo?
