### 2021-09-21 #3642 @kthyng
Presumably the people listed made minor contributions to the software at some point and were auto-included by Zenodo. If on the other hand they were authors of the software, they should be included as co-authors to the paper itself.
