### 2021-09-21 #3642 @solegalli
Thank you!. It's been updated.

We discussed the authorship in the [pre-review PR](https://github.com/openjournals/joss-reviews/issues/3592#issuecomment-895119150) :)
