### 2021-09-22 #3642 @kthyng
@openjournals/dev it's been awhile and the DOI page is still not resolving.. any ideas?
