### 2021-09-24 #3642 @kthyng
@openjournals/dev, @arfon  I'm going to stop checking this after today so I wanted to get one more ping in. The DOI is not resolving for me, 2 days later.
