### 2021-09-27 #3642 @arfon
> @openjournals/dev, @arfon I'm going to stop checking this after today so I wanted to get one more ping in. The DOI is not resolving for me, 2 days later.

I re-deposited the metadata. Not sure what was going on here but the DOI is resolving now.
