### 2021-09-27 #3642 @arfon
AUTHOR – your paper is now accepted and published in JOSS :zap::rocket::boom:
